[wasm] fix
